bank_attrition_project/requirements.txt
pyspark==3.5.0
pymongo==4.6.0
streamlit==1.28.1
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0
